Move interface to it's own class
